add service configuration tests
